+++ inst/tinytest/test-markdown.R
@@ -45,5 +45,16 @@ tab <- tt(test) |>
 expect_snapshot_print(tab, label = "markdown-group_j_wider_2")
 
 
+# Issue #133
+tab <- tt(mtcars[1:8, 1:3]) |>
+  group_tt(i = list("Hello" = 3, "World" = 6)) |>
+  group_tt(j = list("Hello" = 1, "World" = 2:3)) |>
+  style_tt(i = 2:4, italic = TRUE) |>
+  style_tt(i = 6, strikeout = TRUE) |>
+  style_tt(i = 8, bold = TRUE)
+expect_snapshot_print(tab, label = "markdown-group_i_style_tt")
+
+
+
 
 options(tinytable_print_output = NULL)