+++ NEWS.md
@@ -8,6 +8,7 @@ New:
 - Line plots: `plot_tt(fun = "line")`
 - `format_tt(j=NULL, escape=TRUE)` now escapes column headers in addition to all cells.
 - `style_tt()`: `rowspan` and `colspan` arguments are now supported for LaTeX and HTML tables.
+- `style_tt()`: `alignv` argument is now supported for LaTeX and HTML
 
 Bugfix: