+++ NEWS.md
@@ -13,6 +13,7 @@ New:
 Bugfix:
 
 - Markdown group columns when labels are wider than columns. Thanks to @etiennebacher for report #127.
+- Markdown group rows broke indexing when using `style_tt()`. Thanks to @strengejacke for report #133.
 
 ## 0.0.3